requirements.txt
pytorch_lightning>=1.3.1,<=1.4.2
batchgenerators>=0.24
nnunet==1.7.1
scipy
scikit-learn
scikit-image>=0.14
pandas>=0.8.1
PyYAML>=5.1, !=5.4.*
nevergrad
dicom2nifti
medpy
# SimpleITK<2.1.0
tqdm
loguru
hydra-core>=1.1.0
mlflow
GitPython
matplotlib
seaborn
torchmetrics>=0.7.0,<=0.7.3
